requirements.txt
python-dotenv==1.0.0
upstash-redis==1.0.0
python-telegram-bot==20.7
playwright==1.41.0
apscheduler==3.10.4
pydantic==2.5.3
pydantic-settings==2.1.0
